Added tree types and helpers methods.
The idea is to support generic tree types with minimal effort. You can
treat a binary tree as an ordered tree. All of this comes for free if
you use the convience methods which use some of the binary tree methods
(like Left() and Right()) to generate Children(), ChildCount(), and
GetChild().

The interface is read-only, you can't modify the structure of the trees.
I am imagining that you would need to know what type of tree you had
before you started modifying it anyway.