Added support for Marshal/Unmarshal to types
The idea here is to make the simple containers (set, list, etc...)
have marshable versions which depend on their items being marshalable.
So it would look like this:

	type MHashable interface {
		Hashable
		Marshalable
	}

	type MSortedSet SortedSet

	func (m *MSortedSet) Add(item MHashable) error {
		s := (*SortedSet)(m)
		return s.Add(item)
	}

which ensures that all of the items in the container are Marshable. Then
MSortedSet additionally implements Marshalable itself. The trick here is
knowing which Unmarshal to call. This can be done either via the `gob`
package (which could bloat binary size). Or optionally via registering
such a function with the Container. That could look like

	type MSortedSet struct {
		Set *SortedSet
		UnmarshalItem func([]byte) (MHashable, error)
	}

which would allow for more efficient unmarshalling/marshalling when that
function is known while still being able to (potentially) fall back on
the `gob` encoder. The way that would look is the client would register
the inner item with `gob` and the container would use gob to decode it.
This would allow for hetrogenous containers. A final option would be to
forget about `gob` encoding and all of the nonsense and just ensure that
each item conforms to Marshalable (on encode). Still keep the MContainer
versions as the Marshable business should be strictly optional.

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>